set up controller for updatingan order
