+++ server/DB/controllers/orders.js
@@ -1,54 +1,13 @@
 import db from '../db';
 
-const Order = {
-
-
-  async getAll(req, res) {
-    const findAllQuery = 'SELECT * FROM orders';
-    try {
-      const { rows, rowCount } = await db.query(findAllQuery);
-      return res.status(200).send({ rows, rowCount });
-    } catch (error) {
-      return res.status(400).send(error);
-    }
-  },
-
-  async update(req, res) {
-    const findOneQuery = 'SELECT * FROM orders WHERE id=$1';
-    const updateOneQuery = `UPDATE orders
-      SET success=$1,low_point=$2,take_away=$3,modified_date=$4
-      WHERE id=$5 returning *`;
-    try {
-      const { rows } = await db.query(findOneQuery, [req.params.id]);
-      if (!rows[0]) {
-        return res.status(404).send({ message: 'reflection not found' });
-      }
-      const values = [
-        req.body.success || rows[0].success,
-        req.body.low_point || rows[0].low_point,
-        req.body.take_away || rows[0].take_away,
-        new Date(),
-        req.params.id,
-      ];
-      const response = await db.query(updateOneQuery, values);
-      return res.status(200).send(response.rows[0]);
-    } catch (err) {
-      return res.status(400).send(err);
-    }
-  },
-
-  async delete(req, res) {
-    const deleteQuery = 'DELETE FROM orders WHERE id=$1 returning *';
-    try {
-      const { rows } = await db.query(deleteQuery, [req.params.id]);
-      if (!rows[0]) {
-        return res.status(404).send({ message: 'reflection not found' });
-      }
-      return res.status(204).send({ message: 'deleted' });
-    } catch (error) {
-      return res.status(400).send(error);
-    }
-  },
+const orders = (req, res) => {
+  const findAllQuery = 'SELECT * FROM orders';
+  try {
+    const { rows, rowCount } = db.query(findAllQuery);
+    return res.status(200).send({ rows, rowCount });
+  } catch (error) {
+    return res.status(400).send(error);
+  }
 };
 
-export default Order;
+export default orders;
