+++ server/router/api1.js
@@ -4,7 +4,8 @@ import jsOrders from '../data/getOrders';
 import dbOrders from '../DB/controllers/orders';
 import jsNewOrder from '../data/newOrder';
 import dbNewOrder from '../DB/controllers/newOrder';
-import updateOrder from '../data/updateOrder';
+import jsUpdateOrder from '../data/updateOrder';
+import dbUpdateOrder from '../DB/controllers/updateOrder';
 import jsOrder from '../data/order';
 import dbOrder from '../DB/controllers/order';
 
@@ -12,6 +13,7 @@ dotenv.config();
 const order = process.env.TYPE === 'db' ? dbOrder : jsOrder;
 const orders = process.env.TYPE === 'db' ? dbOrders : jsOrders;
 const newOrder = process.env.TYPE === 'db' ? dbNewOrder : jsNewOrder;
+const updateOrder = process.env.TYPE === 'db' ? dbUpdateOrder : jsUpdateOrder;
 
 const api = Router();
 api.use(json());
